x/crypto: implement kbkdf
NIST SP800-108 aka kbkdf is a key-based KDF that uses a keyed PRF (e.g.,
HMAC or CMAC).

This change implements only the HMAC-CTR mode of the KDF, which is the
mode used by Trusted Platform Modules.

SP800-108 suggests that some fields may be omitted or reordered on a
case-by-case basis, but TPM (and openssl) agree on not omitting or
reordering them, and they also agree on a big-endian 32-bit
representation for counters and lengths.

NIST provides some test vectors at
https://csrc.nist.gov/projects/cryptographic-algorithm-validation-program/key-derivation,
but these test vectors are less opinionated about the "fixed input data"
(i.e., everything but the counter) and are not useful for testing the
actual API of a kbkdf implementation. Thus, openssl was used to generate
the test vectors for this implementation.

For simplicity, this change implements kbkdf using pbkdf2 as a model
implementation: return a []byte containing the complete key (rather than
returning an io.Reader). Rather than return an error in the case that
the caller provided nonsensical keyLen, this implementation panics.

The NIST publication can be found at
https://csrc.nist.gov/publications/detail/sp/800-108/final

Change-Id: I56f4cb5f4e13cc80dba74bbcea77718f5b04f31b