ocsp: use produced at timestamp of template when creating a response
Use the template parameter's ProducedAt value for generating response
data via the CreateResponse func when the given value is non-zero. This
provides a way of generating a deterministic ProducedAt value for
responses. When the value is zero, default to the old behavior of
setting the timestamp to the previous minute UTC.

Fixes golang/go#52976

Change-Id: Iedda894536fe88c5e7f11b187a3b110d91563ff7