acme/autocert: fix renewal timer issue
Block when creating the renewal timer, rather than doing it in a
goroutine, forcing the creation to be under the renewalMu lock. This
fixes an issue where startRenew and stopRenew are called very closely
together, and due to lock ordering, stopRenew may be called before
startRenew, resulting in the appearance that the renewal timer has been
stopped before it has actually been created.

In particular this issue can manifest in TestGetCertiifcate sub-tests,
where a httptest server reuses a port across two of the sub-tests. In
this case, the renewal calls end up creating dirty state for the
subsequent test, which can cause confusing behavior (such as attempting
to register an account twice.)

Fixes golang/go#52494

Change-Id: I95420d3fd877572a0b9e408d2f8cd353f6a4e80e