ed25519: make consts an alias for crypto/ed25519
This is a follow-up to Change-Id: I860ed963a0660753d01c89014c21360b239a38ac (commit 20e1d8d225ab8b19a0e7dc00fa5095216b8d6553), which dropped compatibility with go1.12, and made this package an alias / wrapper for crypto/ed25519 in stdlib.

While the above change aliased types and created wrappers for functions, it left out the consts. I looked at the code review and commit message on that change, but did not find a specific reason to leave these out. This patch updates the consts to be an alias as well, which should make it even more transparent that this package is an alias.

Change-Id: I2e2a75dcb8a507451844d25b7afa6a935f5de18b
GitHub-Last-Rev: 070f3a658bdd92974ba315321c18a4198ba9593d
GitHub-Pull-Request: golang/crypto#237