all: use error wrap when possible
I wanted to be able to use error.Is to match io or syscall errors on ssh network error.

This is currently not possible because errors are not wrapped.

This change replace `%v` verb with `%w` in fmt.Errorf() calls.

I was looking to fix this in `ssh` package but found that it could be extended to other package.
This change should be transparent to existing code since they can still rely on old error comparison but will provide more details in case of check with `errors.Is` or `errors.As`.

Change-Id: I38e413f591e3c258fe9cd0a0de561d99dd847bd7
GitHub-Last-Rev: 7527590b58dea491c161aa68953fce34aa75ca76
GitHub-Pull-Request: golang/crypto#253