ssh: expose Connection type
The ssh.NewServerConn and ssh.NewClientConn is declared as returning
an interface type (ssh.Conn). This has been a mistake, as the
connection type is not something that normally has multiple
implementations. On the contrary, we want to be able to expose new
functionality (eg. algorithms used, requesting rekeying, etc.).

To enable this, expose the Connection type, and document that this is
returned from NewClientConn and NewServerConn.

Change-Id: I755251ad79e3ea99ef2bc3b1e98988d8050ba6b6