ssh: add server side multi-step authentication
Add support for sending back partial success to the client while
handling authentication in the server. This is implemented by a special
error that can be returned by any of the authentication methods, which
contains the authentication methods to offer next.

This patch is based on CL 399075 with some minor changes and the addition
of test cases.

Fixes golang/go#61447

Change-Id: I05c8f913bb407d22c2e41c4cbe965e36ab4739b0