Small optimizations of BCrypt + separate version where password hash and secret are separated
**_Changes to bcrypt:_**

- Removed encoding and decoding that happened between generation of unencodedsalt and expensiveblowfishsetup
- CSalt was consequently removed and we now pass salt directly to function
- p.salt is now encoded at the end of newfrompassword function
- Decodecost/version now take pointers to hashed secret and mutate the []array directly in function
- For a few bcrypt_test functions salt was changed to their decoded values due to expensiveblowfishsetup not decoding the salt anymore

Result:

- ns/operation time stays about consistent, maybe very slightly better. But we see roughly 20% less memory allocations/operation.

_**BiiCrypt (Naming convention just to differentiate them, but should be changed to something more concise):**_

The problem it is trying to solve:

Taking the password hash out of the database itself is unnecessary and can lead to vulnerabilities. Micro service architectures with endpoints communicating to remote databases should not have to send passwords over the wire. Doing so would involve allowing password hashes to be accessible/readable to the endpoint's dbuser and transmittable over tcp/ip. In this scenario, if the dbuser is compromised, tables containing sensitive info can be at risk. This can be and should be avoided.

The approach I believe would be best would be to lock down the password hash column (and table at that) to only the root user and not give any other permissions on said table to anyone else. Root user should only be accessible through unix sockets. This would mean that no password can ever be transposed through the wire or be read, unless the root user is compromised. To achieve this password hashes must generated without the original hash. Which it wasn't needed for to begin with, it is only used to compare the two.

If there is something that I am missing or something I am yet to consider please let me know.

**_BiiCrypt modifications_**

- Function returns both password hash and secret separately
- Function generatewithsecret takes user inputed password and stored secret to generate a hash
- STD base64 encoding is used instead of the modified version in bcrypt
- Function Reformat() is included. Pass your old password hash from v2 to have it split up. It returns the password hash and secret. The salt and secret are std base64 encoded.
- notsosecret() derives from hash() function. But only returns the secret now

**_Important notes_**

- Proper thorough testing is yet to be conducted but I have included a few from the original in biicrypt_testing.go.
- Under normal circumstances, this approach requires 2 db query's for all authentication attempts vs the one for failed attempts from the original approach.

Change-Id: Ia0d7d14e6dad375366b46863a696bb3f408a53d6
GitHub-Last-Rev: 2eff55068c5414d9c62a592021480456759bdebf
GitHub-Pull-Request: golang/crypto#271