ed25519: update godoc for consts in crypto/ed25519
This is a follow-up to CL 317169, which dropped compatibility with
go1.12, and made this package an alias / wrapper for crypto/ed25519
in stdlib.

This patch updates their godoc to include a docs-link to the equivalent
consts in stdlib.

Change-Id: I2e2a75dcb8a507451844d25b7afa6a935f5de18b
GitHub-Last-Rev: b307d43f3838f2376e16c24d478ea85df9568dba
GitHub-Pull-Request: golang/crypto#237